refactor(auth): split api.go god object per active-package template (#575 B4)
Pure file moves: api.go (2296 L) → wiring-only (309 L incl. Router and
the SetAuthRateLimiter brute-force-protection setter, wired from
api/base.go) + session_handlers / tenant_handlers / account_handlers /
rbac_handlers / password_handlers / parent_account_handlers + types.go.
Existing mfa/passkey/invitation files untouched. Route table unchanged
(B0 golden).
