+++ backend/api/auth/api.go
@@ -1,37 +1,18 @@
 package auth
 
 import (
-	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"log/slog"
 	"net/http"
-	"slices"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/jwtauth/v5"
 	"github.com/go-chi/render"
-	validation "github.com/go-ozzo/ozzo-validation"
-	"github.com/go-ozzo/ozzo-validation/is"
 	"github.com/uptrace/bun"
 
 	"github.com/moto-nrw/project-phoenix/api/common"
 	"github.com/moto-nrw/project-phoenix/auth/authorize"
 	"github.com/moto-nrw/project-phoenix/auth/jwt"
-	"github.com/moto-nrw/project-phoenix/internal/clientip"
-	"github.com/moto-nrw/project-phoenix/internal/schoolclass"
-	"github.com/moto-nrw/project-phoenix/internal/strutil"
-	authModel "github.com/moto-nrw/project-phoenix/models/auth"
-	configModel "github.com/moto-nrw/project-phoenix/models/config"
-	userModel "github.com/moto-nrw/project-phoenix/models/users"
 	authService "github.com/moto-nrw/project-phoenix/services/auth"
 	configSvc "github.com/moto-nrw/project-phoenix/services/config"
-	"github.com/moto-nrw/project-phoenix/services/parentmessaging"
 	platformSvc "github.com/moto-nrw/project-phoenix/services/platform"
 	usersService "github.com/moto-nrw/project-phoenix/services/users"
 	"github.com/moto-nrw/project-phoenix/tenant"
@@ -39,17 +20,14 @@ import (
 
 // Constants for permission strings, headers, route patterns, and error messages (S1192 - avoid duplicate string literals)
 const (
-	permUsersCreate                    = "users:create"
-	permUsersManage                    = "users:manage"
-	permUsersList                      = "users:list"
-	permRolesRead                      = "roles:read"
-	permUsersUpdate                    = "users:update"
-	permRolesManage                    = "roles:manage"
-	headerUserAgent                    = "User-Agent"
-	pathPermissionID                   = "/{permissionId}"
-	pathPermissions                    = "/permissions"
-	errPasswordsNotMatch               = "passwords do not match"
-	tenantResolveSettingFailureMessage = "tenant resolve setting failed"
+	permUsersCreate  = "users:create"
+	permUsersManage  = "users:manage"
+	permUsersList    = "users:list"
+	permRolesRead    = "roles:read"
+	permUsersUpdate  = "users:update"
+	permRolesManage  = "roles:manage"
+	pathPermissionID = "/{permissionId}"
+	pathPermissions  = "/permissions"
 )
 
 // Resource defines the auth resource
@@ -328,1969 +306,4 @@ func (rs *Resource) Router() chi.Router {
 	})
 
 	return r
-}
-
-// TenantResolveResponse represents the public tenant info returned by resolve
-type TenantResolveResponse struct {
-	TenantID         int64           `json:"tenant_id"`
-	Slug             string          `json:"slug"`
-	Name             string          `json:"name"`
-	Subdomain        string          `json:"subdomain"`
-	OrganizationID   int64           `json:"organization_id"`
-	OrganizationName string          `json:"organization_name"`
-	Hidden           bool            `json:"hidden"`
-	Settings         json.RawMessage `json:"settings"`
-	// PresenceMode is the tenant's resolved operations.presence_mode setting
-	// ("detailed" | "binary"). Included so the frontend can render the right
-	// badge component (PresenceBadge vs LocationBadge) without a second call.
-	// Defaults to "detailed" if SettingsService is unavailable.
-	PresenceMode string `json:"presence_mode"`
-	// StudentPhotosEnabled is the tenant's resolved
-	// operations.student_photos_enabled setting. Surfaced here (rather than
-	// via /api/settings/schema) because every authenticated user — including
-	// non-admin Betreuer who don't carry config:read — needs the flag to
-	// decide whether to render avatars on student cards. Defaults to false
-	// when the setting is missing or unresolvable.
-	StudentPhotosEnabled bool `json:"student_photos_enabled"`
-	// NFCEnabled is the tenant's resolved attendance.nfc_enabled setting.
-	// It is public tenant-shell metadata because non-admin staff also need to
-	// know whether NFC-only areas like classic activities should appear.
-	NFCEnabled bool `json:"nfc_enabled"`
-	// ParentMessagingEnabled is the tenant's resolved operations.parent_notes_enabled
-	// setting. Surfaced here (like the photo flag) so non-admin staff — who don't
-	// carry config:read — can hide the "Neue Nachricht" compose entry points when
-	// the school has parent-OGS messaging turned off, instead of composing into a
-	// 403 dead-end. Defaults to false when the setting is missing/unresolvable.
-	ParentMessagingEnabled bool `json:"parent_messaging_enabled"`
-	// DisplayEnabled is the tenant's resolved display.enabled setting. The
-	// Info-Point Dashboard is opt-in and defaults off, so the frontend needs
-	// this to hide the sidebar entry / admin page for schools that haven't
-	// turned it on. Defaults to false when the setting is missing/unresolvable.
-	DisplayEnabled bool `json:"display_enabled"`
-	// GradeLevelMax is the tenant's enrollment.grade_level_max setting.
-	// Timetable users need it even without config:read, so it belongs to the
-	// same public tenant-shell metadata contract as the feature flags above.
-	// Unlike optional display flags, this constraint fails closed: omitting or
-	// inventing it would make enrollment forms disagree with server validation.
-	GradeLevelMax int `json:"grade_level_max"`
-	// CareOfferingsEnabled is shell metadata for approved-child offering
-	// corrections. Missing metadata stays enabled for compatibility with older
-	// backends; an explicit settings-resolution error fails closed so the UI
-	// cannot offer a mutation that the authoritative service will reject.
-	CareOfferingsEnabled bool   `json:"care_offerings_enabled"`
-	AttendanceWebEnabled bool   `json:"attendance_web_enabled"`
-	GroupMode            string `json:"group_mode"`
-	ShowTimetableCounts  bool   `json:"show_timetable_counts"`
-	WaitlistEnabled      bool   `json:"waitlist_enabled"`
-}
-
-type tenantShellSettings struct {
-	presenceMode           string
-	studentPhotosEnabled   bool
-	nfcEnabled             bool
-	parentMessagingEnabled bool
-	displayEnabled         bool
-	careOfferingsEnabled   bool
-	attendanceWebEnabled   bool
-	groupMode              string
-	showTimetableCounts    bool
-	waitlistEnabled        bool
-}
-
-func resolveTenantGradeLevelMax(
-	ctx context.Context,
-	settings configSvc.SettingsService,
-	tenantID int64,
-) (int, error) {
-	if settings == nil {
-		return 0, errors.New("settings service not configured for grade_level_max")
-	}
-	value, err := settings.ResolveIntForTenant(ctx, tenantID, configModel.KeyEnrollmentGradeLevelMax)
-	if err != nil {
-		return 0, fmt.Errorf("resolve grade_level_max: %w", err)
-	}
-	if value < schoolclass.MinGradeLevel || value > schoolclass.MaxGradeLevel {
-		return 0, fmt.Errorf(
-			"grade_level_max %d outside %d..%d",
-			value,
-			schoolclass.MinGradeLevel,
-			schoolclass.MaxGradeLevel,
-		)
-	}
-	return value, nil
-}
-
-// resolveTenant handles GET /auth/tenant/resolve?slug={slug}
-func (rs *Resource) resolveTenant(w http.ResponseWriter, r *http.Request) {
-	slug := strings.TrimSpace(r.URL.Query().Get("slug"))
-	if slug == "" {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("slug query parameter is required")))
-		return
-	}
-
-	school, err := rs.SchoolService.GetSchoolBySubdomain(r.Context(), slug)
-	if err != nil || school == nil || school.IsDeleted() || !school.Active {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("tenant not found")))
-		return
-	}
-
-	// Parse settings JSON; fall back to empty object on invalid data
-	settings := json.RawMessage(school.Settings)
-	if !json.Valid(settings) {
-		settings = json.RawMessage(`{}`)
-	}
-
-	var orgName string
-	if school.Organization != nil {
-		orgName = school.Organization.Name
-	}
-
-	// GradeLevelMax is a validation constraint, not an optional display hint.
-	// The registry returns the legitimate default for tenants without an
-	// override; a missing settings service, read error, or corrupt value must
-	// fail the whole resolve contract instead of silently substituting grade 4.
-	gradeLevelMax, err := resolveTenantGradeLevelMax(r.Context(), rs.SettingsService, school.ID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServerWrap(
-			http.StatusText(http.StatusInternalServerError),
-			fmt.Errorf("tenant resolve: %w", err),
-		))
-		return
-	}
-	resolved := rs.resolveTenantShellSettings(r.Context(), school.ID)
-
-	resp := &TenantResolveResponse{
-		TenantID:               school.ID,
-		Slug:                   school.Slug,
-		Name:                   school.Name,
-		Subdomain:              school.Subdomain,
-		OrganizationID:         school.OrganizationID,
-		OrganizationName:       orgName,
-		Hidden:                 school.Hidden,
-		Settings:               settings,
-		PresenceMode:           resolved.presenceMode,
-		StudentPhotosEnabled:   resolved.studentPhotosEnabled,
-		NFCEnabled:             resolved.nfcEnabled,
-		ParentMessagingEnabled: resolved.parentMessagingEnabled,
-		DisplayEnabled:         resolved.displayEnabled,
-		GradeLevelMax:          gradeLevelMax,
-		CareOfferingsEnabled:   resolved.careOfferingsEnabled,
-		AttendanceWebEnabled:   resolved.attendanceWebEnabled,
-		GroupMode:              resolved.groupMode,
-		ShowTimetableCounts:    resolved.showTimetableCounts,
-		WaitlistEnabled:        resolved.waitlistEnabled,
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Tenant resolved successfully")
-}
-
-func (rs *Resource) resolveTenantShellSettings(ctx context.Context, tenantID int64) tenantShellSettings {
-	resolved := tenantShellSettings{
-		presenceMode:         configSvc.ResolvePresenceModeForTenant(ctx, rs.SettingsService, tenantID, nil),
-		careOfferingsEnabled: true,
-		groupMode:            configModel.GroupModeFixedGroups,
-		showTimetableCounts:  true,
-		waitlistEnabled:      true,
-	}
-	if rs.SettingsService == nil {
-		return resolved
-	}
-
-	if value, err := rs.SettingsService.ResolveStringForTenant(ctx, tenantID, configModel.KeyStudentPhotosEnabled); err == nil {
-		resolved.studentPhotosEnabled = value == "true"
-	}
-	if value, err := rs.SettingsService.ResolveBoolForTenant(ctx, tenantID, configModel.KeyAttendanceNFCEnabled); err == nil {
-		resolved.nfcEnabled = value
-	}
-	if value, err := rs.SettingsService.ResolveBoolForTenant(ctx, tenantID, configModel.KeyDisplayEnabled); err == nil {
-		resolved.displayEnabled = value
-	}
-	resolved.careOfferingsEnabled = rs.resolveTenantShellBool(ctx, tenantID, configModel.KeyEnrollmentCareOfferingsEnabled, false, slog.LevelError)
-	resolved.attendanceWebEnabled = rs.resolveTenantShellBool(ctx, tenantID, configModel.KeyAttendanceWebEnabled, false, slog.LevelError)
-	resolved.showTimetableCounts = rs.resolveTenantShellBool(ctx, tenantID, configModel.KeyTimetableShowExpectedChildrenCount, true, slog.LevelWarn)
-	resolved.waitlistEnabled = rs.resolveTenantShellBool(ctx, tenantID, configModel.KeyEnrollmentWaitlistEnabled, true, slog.LevelError)
-	resolved.groupMode = rs.resolveTenantGroupMode(ctx, tenantID)
-
-	// Messaging compose visibility intentionally fails open so it stays in
-	// lockstep with the unread badge, inbox row pills, and reply path.
-	resolved.parentMessagingEnabled = parentmessaging.MessagingEnabledForTenant(ctx, rs.SettingsService, tenantID, nil)
-	return resolved
-}
-
-func (rs *Resource) resolveTenantShellBool(ctx context.Context, tenantID int64, key string, fallback bool, level slog.Level) bool {
-	value, err := rs.SettingsService.ResolveBoolForTenant(ctx, tenantID, key)
-	if err == nil {
-		return value
-	}
-	logTenantResolveSettingFailure(ctx, tenantID, key, err, level)
-	return fallback
-}
-
-func (rs *Resource) resolveTenantGroupMode(ctx context.Context, tenantID int64) string {
-	value, err := rs.SettingsService.ResolveStringForTenant(ctx, tenantID, configModel.KeyGroupMode)
-	if err != nil {
-		logTenantResolveSettingFailure(ctx, tenantID, configModel.KeyGroupMode, err, slog.LevelError)
-		return configModel.GroupModeFixedGroups
-	}
-	if value == configModel.GroupModeOpenCare {
-		return value
-	}
-	return configModel.GroupModeFixedGroups
-}
-
-func logTenantResolveSettingFailure(ctx context.Context, tenantID int64, key string, err error, level slog.Level) {
-	slog.Default().LogAttrs(ctx, level, tenantResolveSettingFailureMessage,
-		slog.Int64("tenant_id", tenantID),
-		slog.String("key", key),
-		slog.String("error", err.Error()))
-}
-
-// SwitchTenantRequest represents the switch-tenant request payload
-type SwitchTenantRequest struct {
-	TenantSlug string `json:"tenant_slug"`
-}
-
-// Bind validates the switch-tenant request
-func (req *SwitchTenantRequest) Bind(_ *http.Request) error {
-	req.TenantSlug = strings.TrimSpace(req.TenantSlug)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.TenantSlug, validation.Required),
-	)
-}
-
-// switchTenant handles POST /auth/switch-tenant
-func (rs *Resource) switchTenant(w http.ResponseWriter, r *http.Request) {
-	req := &SwitchTenantRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	// Get account ID from JWT claims
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	accessToken, refreshToken, err := rs.AuthService.SwitchTenant(r.Context(), int64(claims.ID), req.TenantSlug)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountNotFound))
-			case errors.Is(authErr.Err, authService.ErrAccountInactive):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountInactive))
-			case errors.Is(authErr.Err, authService.ErrTenantNotFound):
-				common.RenderError(w, r, common.ErrorNotFound(authService.ErrTenantNotFound))
-			case errors.Is(authErr.Err, authService.ErrTenantAccessDenied):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrTenantAccessDenied))
-			default:
-				common.RenderError(w, r, common.ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	// Return tokens in the same format as login
-	render.JSON(w, r, TokenResponse{
-		AccessToken:  accessToken,
-		RefreshToken: refreshToken,
-	})
-}
-
-// AccountTenantResponse represents a tenant available to the authenticated user
-type AccountTenantResponse struct {
-	TenantID         int64  `json:"tenant_id"`
-	Slug             string `json:"slug"`
-	Name             string `json:"name"`
-	Subdomain        string `json:"subdomain"`
-	OrganizationID   int64  `json:"organization_id"`
-	OrganizationName string `json:"organization_name"`
-}
-
-// listAccountTenants handles GET /auth/account/tenants
-func (rs *Resource) listAccountTenants(w http.ResponseWriter, r *http.Request) {
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	schools, err := rs.SchoolService.ListActiveSchoolsByAccountID(r.Context(), int64(claims.ID))
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]AccountTenantResponse, 0, len(schools))
-	for _, school := range schools {
-		var orgName string
-		if school.Organization != nil {
-			orgName = school.Organization.Name
-		}
-		responses = append(responses, AccountTenantResponse{
-			TenantID:         school.ID,
-			Slug:             school.Slug,
-			Name:             school.Name,
-			Subdomain:        school.Subdomain,
-			OrganizationID:   school.OrganizationID,
-			OrganizationName: orgName,
-		})
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account tenants retrieved successfully")
-}
-
-// PublicTenantResponse is the public-facing tenant info returned by the
-// unauthenticated /auth/tenants endpoint. It intentionally omits internal
-// database IDs (tenant_id, organization_id) to prevent enumeration.
-type PublicTenantResponse struct {
-	Slug             string `json:"slug"`
-	Name             string `json:"name"`
-	Subdomain        string `json:"subdomain"`
-	OrganizationName string `json:"organization_name"`
-}
-
-// listTenants handles GET /auth/tenants (public, no auth required).
-// Uses ListPublic to exclude hidden schools from the public landing page.
-// Hidden schools remain accessible via direct subdomain link.
-func (rs *Resource) listTenants(w http.ResponseWriter, r *http.Request) {
-	schools, err := rs.SchoolService.ListPublicSchools(r.Context())
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]PublicTenantResponse, 0, len(schools))
-	for _, school := range schools {
-		var orgName string
-		if school.Organization != nil {
-			orgName = school.Organization.Name
-		}
-		responses = append(responses, PublicTenantResponse{
-			Slug:             school.Slug,
-			Name:             school.Name,
-			Subdomain:        school.Subdomain,
-			OrganizationName: orgName,
-		})
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Tenants retrieved successfully")
-}
-
-// LoginRequest represents the login request payload
-type LoginRequest struct {
-	Email      string `json:"email"`
-	Password   string `json:"password"`
-	TenantSlug string `json:"tenant_slug,omitempty"`
-}
-
-// Bind validates the login request
-func (req *LoginRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Password, validation.Required),
-	)
-}
-
-// TokenResponse represents the token response payload
-type TokenResponse struct {
-	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
-}
-
-// LoginResponse is the discriminated response shape /auth/login returns now
-// that MFA can short-circuit the token pair. The `status` field tells the
-// client which fields to read:
-//
-//   - status == "authenticated"           → access_token + refresh_token
-//     populated; the frontend stores them as before.
-//   - status == "mfa_required"             → challenge_token + masked_email
-//     populated; the frontend renders the code-entry UI and POSTs to
-//     /auth/mfa/verify.
-//   - status == "mfa_enrollment_required"  → access_token carries a
-//     short-lived enrollment-scoped JWT (no refresh_token). The frontend
-//     uses it as the Authorization header for /auth/mfa/enroll/* and
-//     replaces it with the real token pair returned by
-//     /auth/mfa/enroll/confirm. The token is REJECTED by the regular
-//     Authenticator middleware, so it cannot be used to bypass MFA.
-//
-// mfa_enrollment_required is retained as a legacy hint for clients that key
-// off the boolean rather than the status discriminator; new code should
-// branch on `status`.
-type LoginResponse struct {
-	Status                string `json:"status"`
-	AccessToken           string `json:"access_token,omitempty"`
-	RefreshToken          string `json:"refresh_token,omitempty"`
-	ChallengeToken        string `json:"challenge_token,omitempty"`
-	MaskedEmail           string `json:"masked_email,omitempty"`
-	MFAEnrollmentRequired bool   `json:"mfa_enrollment_required,omitempty"`
-	// TrustedDeviceEnabled is sent on the mfa_required branch so the
-	// frontend can hide the "remember this device" checkbox when the
-	// tenant admin has disabled the feature. A missing field defaults to
-	// "enabled" on the client for backwards compatibility.
-	TrustedDeviceEnabled *bool `json:"trusted_device_enabled,omitempty"`
-	// TrustedDeviceDays mirrors security.mfa_trusted_device_days for the
-	// tenant. The frontend uses it to render the dynamic "Auf diesem
-	// Gerät N Tage merken" label so the checkbox always reflects the
-	// cookie lifetime the backend will actually issue.
-	TrustedDeviceDays *int `json:"trusted_device_days,omitempty"`
-}
-
-// login handles user login. The handler is a thin orchestrator: it pulls
-// the trusted-device cookie off the request, calls LoginWithMFAGate, and
-// translates the discriminated LoginResult into a JSON shape the frontend
-// can branch on.
-func (rs *Resource) login(w http.ResponseWriter, r *http.Request) {
-	req := &LoginRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	// Pull the trusted-device cookie if the browser sent one. The MFA gate
-	// uses it to skip the second factor for users on a previously-marked
-	// device. Empty / missing cookie is fine — the service is nil-tolerant.
-	var trustedDeviceCookie string
-	if c, err := r.Cookie(trustedDeviceCookieName); err == nil {
-		trustedDeviceCookie = c.Value
-	}
-
-	result, err := rs.AuthService.LoginWithMFAGate(
-		r.Context(), req.Email, req.Password, ipAddress, userAgent, req.TenantSlug, trustedDeviceCookie,
-	)
-	if err != nil {
-		rs.handleLoginError(w, r, err)
-		return
-	}
-
-	switch result.Status {
-	case authService.LoginStatusMFARequired:
-		tde := result.TrustedDeviceEnabled
-		tdd := result.TrustedDeviceDays
-		render.JSON(w, r, LoginResponse{
-			Status:               string(authService.LoginStatusMFARequired),
-			ChallengeToken:       result.ChallengeToken,
-			MaskedEmail:          result.MaskedEmail,
-			TrustedDeviceEnabled: &tde,
-			TrustedDeviceDays:    &tdd,
-		})
-		return
-	case authService.LoginStatusMFAEnrollmentRequired:
-		render.JSON(w, r, LoginResponse{
-			Status:                string(authService.LoginStatusMFAEnrollmentRequired),
-			AccessToken:           result.AccessToken,
-			MaskedEmail:           result.MaskedEmail,
-			MFAEnrollmentRequired: true,
-		})
-		return
-	}
-
-	render.JSON(w, r, LoginResponse{
-		Status:       string(authService.LoginStatusAuthenticated),
-		AccessToken:  result.AccessToken,
-		RefreshToken: result.RefreshToken,
-	})
-}
-
-// handleLoginError centralises the error-to-HTTP mapping for /auth/login so
-// both the legacy and the MFA-aware paths agree on what comes back.
-func (rs *Resource) handleLoginError(w http.ResponseWriter, r *http.Request, err error) {
-	var authErr *authService.AuthError
-	if errors.As(err, &authErr) {
-		switch {
-		case errors.Is(authErr.Err, authService.ErrInvalidCredentials):
-			common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrInvalidCredentials))
-		case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-			// Mask the specific error so attackers can't enumerate accounts.
-			common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrInvalidCredentials))
-		case errors.Is(authErr.Err, authService.ErrAccountInactive):
-			common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountInactive))
-		case errors.Is(authErr.Err, authService.ErrTenantNotFound):
-			common.RenderError(w, r, common.ErrorNotFound(authService.ErrTenantNotFound))
-		case errors.Is(authErr.Err, authService.ErrTenantAccessDenied):
-			common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrTenantAccessDenied))
-		case errors.Is(authErr.Err, authService.ErrParentMustUseParentPortal):
-			common.RenderError(w, r, common.ErrorForbidden(authService.ErrParentMustUseParentPortal))
-		case errors.Is(authErr.Err, authService.ErrMFARateLimited):
-			// MFA challenge initiation tripped the 3/15min sliding-window
-			// cap. Surface as 429 so the frontend shows the dedicated "too
-			// many code requests" message instead of a generic 5xx.
-			common.RenderError(w, r, common.ErrorTooManyRequests(authErr.Err))
-		case errors.Is(authErr.Err, authService.ErrMFALocked):
-			// Account hit the failed-attempt lockout threshold while we were
-			// preparing the next challenge. Same HTTP status as rate limit,
-			// distinct message body — handled separately on the frontend.
-			common.RenderError(w, r, common.ErrorTooManyRequests(authErr.Err))
-		case errors.Is(authErr.Err, authService.ErrMFAStatusUnavailable):
-			// MFA gate couldn't determine required/enrolled status (settings
-			// or credentials lookup failed with a non-not-found error).
-			// Refuse this login rather than fail-open. 503 lets the client
-			// retry — the frontend renders it as "Bitte versuche es erneut".
-			common.RenderError(w, r, common.ErrorServiceUnavailable(authErr.Err))
-		default:
-			common.RenderError(w, r, common.ErrorInternalServer(err))
-		}
-		return
-	}
-	common.RenderError(w, r, common.ErrorInternalServer(err))
-}
-
-// RegisterRequest represents the register request payload
-type RegisterRequest struct {
-	Email           string `json:"email"`
-	Username        string `json:"username"`
-	Password        string `json:"password"`
-	ConfirmPassword string `json:"confirm_password"`
-	RoleID          *int64 `json:"role_id,omitempty"` // Optional role assignment
-}
-
-// Bind validates the register request
-func (req *RegisterRequest) Bind(_ *http.Request) error {
-	return validateAccountCredentialFields(req, &req.Email, &req.Username, &req.Password, &req.ConfirmPassword)
-}
-
-// validateAccountCredentialFields normalizes and validates the shared
-// email/username/password(+confirmation) fields of the account-creation
-// payloads (staff registration and parent accounts).
-func validateAccountCredentialFields(structPtr any, email, username, password, confirmPassword *string) error {
-	*email = strings.TrimSpace(strings.ToLower(*email))
-	*username = strings.TrimSpace(*username)
-
-	return validation.ValidateStruct(structPtr,
-		validation.Field(email, validation.Required, is.Email),
-		validation.Field(username, validation.Required, validation.Length(3, 30)),
-		validation.Field(password, validation.Required, validation.Length(8, 0)),
-		validation.Field(confirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if *password != *confirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// AccountResponse represents the account response payload
-type AccountResponse struct {
-	ID          int64    `json:"id"`
-	Email       string   `json:"email"`
-	Username    string   `json:"username,omitempty"`
-	Active      bool     `json:"active"`
-	Roles       []string `json:"roles,omitempty"`
-	Permissions []string `json:"permissions,omitempty"`
-}
-
-// register handles user registration
-func (rs *Resource) register(w http.ResponseWriter, r *http.Request) {
-	req := &RegisterRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	// Authorize role assignment (if role_id specified)
-	roleID, callerTenantID, shouldReturn := rs.authorizeRoleAssignment(w, r, req.RoleID)
-	if shouldReturn {
-		return
-	}
-
-	account, err := rs.AuthService.Register(r.Context(), req.Email, req.Username, req.Password, roleID, callerTenantID)
-	if err != nil {
-		rs.handleRegistrationError(w, r, err)
-		return
-	}
-
-	resp := buildAccountResponse(account)
-	common.Respond(w, r, http.StatusCreated, resp, "Account registered successfully")
-}
-
-// LinkToTenantRequest represents a request to link an existing account to the current tenant.
-type LinkToTenantRequest struct {
-	Email  string `json:"email"`
-	RoleID *int64 `json:"role_id,omitempty"`
-}
-
-func (req *LinkToTenantRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-	)
-}
-
-// linkToTenant links an existing account to the caller's tenant.
-// Requires admin authentication with a valid tenant context.
-func (rs *Resource) linkToTenant(w http.ResponseWriter, r *http.Request) {
-	req := &LinkToTenantRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	// Require admin auth and resolve role + tenant from JWT
-	roleID, callerTenantID, shouldReturn := rs.authorizeRoleAssignment(w, r, req.RoleID)
-	if shouldReturn {
-		return
-	}
-
-	account, err := rs.AuthService.LinkAccountToTenant(r.Context(), req.Email, roleID, callerTenantID)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, common.ErrorNotFound(authErr.Err))
-			case errors.Is(authErr.Err, authService.ErrAccountInactive):
-				common.RenderError(w, r, common.ErrorConflict(authErr.Err))
-			default:
-				common.RenderError(w, r, common.ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	// Return ONLY id and email — never leak roles, username, or active status from other tenants
-	common.Respond(w, r, http.StatusOK, map[string]any{
-		"id":    account.ID,
-		"email": account.Email,
-	}, "Account linked to tenant successfully")
-}
-
-// authorizeRoleAssignment validates the role_id from the request and returns it along with the
-// caller's tenant ID. Auth and permission checks are handled by middleware (Authenticator +
-// TenantMiddleware + RequiresPermission). Returns the role ID, tenant ID, and whether the
-// handler should return early (true = error rendered).
-func (rs *Resource) authorizeRoleAssignment(w http.ResponseWriter, r *http.Request, requestedRoleID *int64) (*int64, int64, bool) {
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	if requestedRoleID == nil || *requestedRoleID <= 0 {
-		common.RenderError(w, r, common.ErrorInvalidRequest(
-			errors.New("role_id is required when creating accounts")))
-		return nil, 0, true
-	}
-
-	if claims.TenantID <= 0 {
-		common.RenderError(w, r, common.ErrorInvalidRequest(
-			authService.ErrTenantRequiredForRoleAssignment))
-		return nil, 0, true
-	}
-
-	// Verify the role exists — TenantTxMiddleware ensures RLS sees tenant-scoped roles
-	if _, err := rs.AuthService.GetRoleByID(r.Context(), int(*requestedRoleID)); err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			common.RenderError(w, r, common.ErrorInvalidRequest(
-				errors.New("specified role does not exist")))
-		} else {
-			slog.Default().Error("role lookup failed",
-				"role_id", *requestedRoleID,
-				"error", err,
-			)
-			common.RenderError(w, r, common.ErrorInternalServerWrap("failed to verify role", err))
-		}
-		return nil, 0, true
-	}
-
-	return requestedRoleID, claims.TenantID, false
-}
-
-// handleRegistrationError handles authentication errors during registration
-func (rs *Resource) handleRegistrationError(w http.ResponseWriter, r *http.Request, err error) {
-	var authErr *authService.AuthError
-	if !errors.As(err, &authErr) {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	switch {
-	case errors.Is(authErr.Err, authService.ErrEmailAlreadyExists):
-		common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrEmailAlreadyExists))
-	case errors.Is(authErr.Err, authService.ErrUsernameAlreadyExists):
-		common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrUsernameAlreadyExists))
-	case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-		common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-	case errors.Is(authErr.Err, authService.ErrTenantRequiredForRoleAssignment):
-		common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrTenantRequiredForRoleAssignment))
-	default:
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-	}
-}
-
-// buildAccountResponse constructs an AccountResponse from an Account model
-func buildAccountResponse(account *authModel.Account) *AccountResponse {
-	resp := &AccountResponse{
-		ID:     account.ID,
-		Email:  account.Email,
-		Active: account.Active,
-	}
-
-	if account.Username != nil {
-		resp.Username = *account.Username
-	}
-
-	roleNames := make([]string, 0, len(account.Roles))
-	for _, role := range account.Roles {
-		roleNames = append(roleNames, role.Name)
-	}
-	resp.Roles = roleNames
-
-	return resp
-}
-
-// refreshToken handles token refresh
-func (rs *Resource) refreshToken(w http.ResponseWriter, r *http.Request) {
-	// Get refresh token from context
-	refreshToken := jwt.RefreshTokenFromCtx(r.Context())
-
-	// Get IP address and user agent for audit logging
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	accessToken, newRefreshToken, err := rs.AuthService.RefreshTokenWithAudit(r.Context(), refreshToken, ipAddress, userAgent)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidToken):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrInvalidToken))
-			case errors.Is(authErr.Err, authService.ErrTokenExpired):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrTokenExpired))
-			case errors.Is(authErr.Err, authService.ErrTokenNotFound):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrTokenNotFound))
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountNotFound))
-			case errors.Is(authErr.Err, authService.ErrAccountInactive):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountInactive))
-			case errors.Is(authErr.Err, authService.ErrTenantNotFound):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrTenantNotFound))
-			default:
-				common.RenderError(w, r, common.ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	// Special case for token refresh endpoint - frontend expects direct token response
-	render.JSON(w, r, TokenResponse{
-		AccessToken:  accessToken,
-		RefreshToken: newRefreshToken,
-	})
-}
-
-// logout handles user logout
-func (rs *Resource) logout(w http.ResponseWriter, r *http.Request) {
-	// Get refresh token from context
-	refreshToken := jwt.RefreshTokenFromCtx(r.Context())
-
-	// Get IP address and user agent for audit logging
-	ipAddress := getClientIP(r)
-	userAgent := r.Header.Get(headerUserAgent)
-
-	err := rs.AuthService.LogoutWithAudit(r.Context(), refreshToken, ipAddress, userAgent)
-	if err != nil {
-		// Even if there's an error, we want to consider the logout successful from the client's perspective
-		// Log the error on the server side for debugging
-		slog.Default().WarnContext(r.Context(), "Logout audit logging failed (client logout still successful)",
-			slog.String("ip", ipAddress),
-			slog.String("error", err.Error()),
-		)
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// ChangePasswordRequest represents the change password request payload
-type ChangePasswordRequest struct {
-	CurrentPassword string `json:"current_password"`
-	NewPassword     string `json:"new_password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the change password request
-func (req *ChangePasswordRequest) Bind(_ *http.Request) error {
-	return validation.ValidateStruct(req,
-		validation.Field(&req.CurrentPassword, validation.Required),
-		validation.Field(&req.NewPassword, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.NewPassword != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// changePassword handles password change
-func (rs *Resource) changePassword(w http.ResponseWriter, r *http.Request) {
-	req := &ChangePasswordRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	// Get user ID from JWT claims
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	err := rs.AuthService.ChangePassword(r.Context(), claims.ID, req.CurrentPassword, req.NewPassword)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidCredentials):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrInvalidCredentials))
-			case errors.Is(authErr.Err, authService.ErrAccountNotFound):
-				common.RenderError(w, r, common.ErrorUnauthorized(authService.ErrAccountNotFound))
-			case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-				common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-			default:
-				common.RenderError(w, r, common.ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// getAccount returns the current user's account details
-func (rs *Resource) getAccount(w http.ResponseWriter, r *http.Request) {
-	// Get user ID and permissions from JWT claims
-	claims := jwt.ClaimsFromCtx(r.Context())
-
-	account, err := rs.AuthService.GetAccountByID(r.Context(), claims.ID)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			if errors.Is(authErr.Err, authService.ErrAccountNotFound) {
-				common.RenderError(w, r, common.ErrorNotFound(authService.ErrAccountNotFound))
-				return
-			}
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	// Convert account to response
-	resp := &AccountResponse{
-		ID:     account.ID,
-		Email:  account.Email,
-		Active: account.Active,
-	}
-
-	if account.Username != nil {
-		resp.Username = *account.Username
-	}
-
-	roleNames := make([]string, 0, len(account.Roles))
-	for _, role := range account.Roles {
-		roleNames = append(roleNames, role.Name)
-	}
-	resp.Roles = roleNames
-
-	// Include permissions from JWT claims
-	resp.Permissions = claims.Permissions
-
-	common.Respond(w, r, http.StatusOK, resp, "Account information retrieved successfully")
-}
-
-// Role Management Request/Response Types
-
-// CreateRoleRequest represents the create role request payload
-type CreateRoleRequest struct {
-	Name        string  `json:"name"`
-	Description string  `json:"description"`
-	BaseRole    *string `json:"base_role,omitempty"`
-}
-
-// Bind validates the create role request
-func (req *CreateRoleRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-	// Empty/blank base_role trims to nil, treated as "field not sent" (preserve
-	// existing value): once set, a base_role cannot be cleared back to NULL via the API.
-	req.BaseRole = strutil.TrimPtrToNil(req.BaseRole)
-
-	if err := validateBaseRole(req.BaseRole); err != nil {
-		return err
-	}
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-	)
-}
-
-// UpdateRoleRequest represents the update role request payload
-type UpdateRoleRequest struct {
-	Name        string  `json:"name"`
-	Description string  `json:"description"`
-	BaseRole    *string `json:"base_role,omitempty"`
-}
-
-// Bind validates the update role request.
-// BaseRole is optional on update — if omitted, the handler preserves the existing value.
-func (req *UpdateRoleRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-	req.BaseRole = strutil.TrimPtrToNil(req.BaseRole)
-
-	// Only validate base_role if the caller explicitly sent one.
-	if req.BaseRole != nil {
-		if err := validateBaseRoleValue(*req.BaseRole); err != nil {
-			return err
-		}
-	}
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-	)
-}
-
-// validateBaseRoleValue checks that a non-nil base_role value is one of the allowed system roles.
-func validateBaseRoleValue(value string) error {
-	if !slices.Contains(authModel.ValidBaseRoles(), value) {
-		return fmt.Errorf("base_role must be one of: %v", authModel.ValidBaseRoles())
-	}
-	return nil
-}
-
-// validateBaseRole checks that base_role is provided and is one of the allowed system roles.
-// This runs at the request layer to return 400 (not 500) for invalid payloads.
-func validateBaseRole(br *string) error {
-	if br == nil {
-		return fmt.Errorf("base_role is required; must be one of: %v", authModel.ValidBaseRoles())
-	}
-	return validateBaseRoleValue(*br)
-}
-
-// RoleResponse represents a role response
-type RoleResponse struct {
-	ID          int64    `json:"id"`
-	Name        string   `json:"name"`
-	Description string   `json:"description"`
-	IsSystem    bool     `json:"is_system"`
-	BaseRole    *string  `json:"base_role,omitempty"`
-	CreatedAt   string   `json:"created_at"`
-	UpdatedAt   string   `json:"updated_at"`
-	Permissions []string `json:"permissions,omitempty"`
-}
-
-// toRoleResponse maps a domain Role to its API response representation.
-func toRoleResponse(role *authModel.Role) *RoleResponse {
-	return &RoleResponse{
-		ID:          role.ID,
-		Name:        role.Name,
-		Description: role.Description,
-		IsSystem:    role.IsSystem,
-		BaseRole:    role.BaseRole,
-		CreatedAt:   role.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   role.UpdatedAt.Format(time.RFC3339),
-	}
-}
-
-// Permission Management Request/Response Types
-
-// CreatePermissionRequest represents the create permission request payload
-type CreatePermissionRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Resource    string `json:"resource"`
-	Action      string `json:"action"`
-}
-
-// Bind validates the create permission request
-func (req *CreatePermissionRequest) Bind(_ *http.Request) error {
-	req.Name = strings.TrimSpace(req.Name)
-	req.Description = strings.TrimSpace(req.Description)
-	req.Resource = strings.TrimSpace(strings.ToLower(req.Resource))
-	req.Action = strings.TrimSpace(strings.ToLower(req.Action))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&req.Description, validation.Length(0, 500)),
-		validation.Field(&req.Resource, validation.Required, validation.Length(1, 50)),
-		validation.Field(&req.Action, validation.Required, validation.Length(1, 50)),
-	)
-}
-
-// UpdatePermissionRequest represents the update permission request payload.
-// Field set, JSON tags, and Bind validation are identical to the create
-// payload, so it is a type alias.
-type UpdatePermissionRequest = CreatePermissionRequest
-
-// PermissionResponse represents a permission response
-type PermissionResponse struct {
-	ID          int64  `json:"id"`
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Resource    string `json:"resource"`
-	Action      string `json:"action"`
-	CreatedAt   string `json:"created_at"`
-	UpdatedAt   string `json:"updated_at"`
-}
-
-// UpdateAccountRequest represents the update account request payload
-type UpdateAccountRequest struct {
-	Email    string `json:"email"`
-	Username string `json:"username,omitempty"`
-}
-
-type EnableCaregiverCapabilityRequest struct {
-	FirstName string `json:"first_name,omitempty"`
-	LastName  string `json:"last_name,omitempty"`
-	Position  string `json:"position,omitempty"`
-}
-
-// Bind validates the update account request
-func (req *UpdateAccountRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-	req.Username = strings.TrimSpace(req.Username)
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-		validation.Field(&req.Username, validation.Length(3, 30)),
-	)
-}
-
-func (req *EnableCaregiverCapabilityRequest) Bind(_ *http.Request) error {
-	req.FirstName = strings.TrimSpace(req.FirstName)
-	req.LastName = strings.TrimSpace(req.LastName)
-	req.Position = strings.TrimSpace(req.Position)
-	return nil
-}
-
-// PasswordResetRequest represents the password reset request payload
-type PasswordResetRequest struct {
-	Email string `json:"email"`
-}
-
-// Bind validates the password reset request
-func (req *PasswordResetRequest) Bind(_ *http.Request) error {
-	req.Email = strings.TrimSpace(strings.ToLower(req.Email))
-
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Email, validation.Required, is.Email),
-	)
-}
-
-// PasswordResetConfirmRequest represents the password reset confirm request payload
-type PasswordResetConfirmRequest struct {
-	Token           string `json:"token"`
-	NewPassword     string `json:"new_password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the password reset confirm request
-func (req *PasswordResetConfirmRequest) Bind(_ *http.Request) error {
-	return validation.ValidateStruct(req,
-		validation.Field(&req.Token, validation.Required),
-		validation.Field(&req.NewPassword, validation.Required, validation.Length(8, 0)),
-		validation.Field(&req.ConfirmPassword, validation.Required, validation.By(func(value interface{}) error {
-			if req.NewPassword != req.ConfirmPassword {
-				return errors.New(errPasswordsNotMatch)
-			}
-			return nil
-		})),
-	)
-}
-
-// CreateParentAccountRequest represents the create parent account request payload
-type CreateParentAccountRequest struct {
-	Email           string `json:"email"`
-	Username        string `json:"username"`
-	Password        string `json:"password"`
-	ConfirmPassword string `json:"confirm_password"`
-}
-
-// Bind validates the create parent account request
-func (req *CreateParentAccountRequest) Bind(_ *http.Request) error {
-	return validateAccountCredentialFields(req, &req.Email, &req.Username, &req.Password, &req.ConfirmPassword)
-}
-
-// ParentAccountResponse represents a parent account response
-type ParentAccountResponse struct {
-	ID        int64  `json:"id"`
-	Email     string `json:"email"`
-	Username  string `json:"username,omitempty"`
-	Active    bool   `json:"active"`
-	CreatedAt string `json:"created_at"`
-	UpdatedAt string `json:"updated_at"`
-}
-
-// Role Management Endpoints
-
-// createRole handles creating a new role
-func (rs *Resource) createRole(w http.ResponseWriter, r *http.Request) {
-	req := &CreateRoleRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	role, err := rs.AuthService.CreateRole(r.Context(), req.Name, req.Description, req.BaseRole)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.Respond(w, r, http.StatusCreated, toRoleResponse(role), "Role created successfully")
-}
-
-// getRoleByID handles getting a role by ID
-func (rs *Resource) getRoleByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	role, err := rs.AuthService.GetRoleByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("role not found")))
-		return
-	}
-
-	// Get permissions for the role
-	permissions, _ := rs.AuthService.GetRolePermissions(r.Context(), id)
-	permissionNames := make([]string, 0, len(permissions))
-	for _, perm := range permissions {
-		permissionNames = append(permissionNames, perm.GetFullName())
-	}
-
-	resp := toRoleResponse(role)
-	resp.Permissions = permissionNames
-
-	common.Respond(w, r, http.StatusOK, resp, "Role retrieved successfully")
-}
-
-// updateRole handles updating a role
-func (rs *Resource) updateRole(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateRoleRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	role, err := rs.AuthService.GetRoleByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("role not found")))
-		return
-	}
-
-	role.Name = req.Name
-	role.Description = req.Description
-	// Preserve existing base_role when the caller omits the field.
-	// System roles must never carry a base_role — ignore silently.
-	if req.BaseRole != nil && !role.IsSystem {
-		role.BaseRole = req.BaseRole
-	}
-
-	if err := rs.AuthService.UpdateRole(r.Context(), role); err != nil {
-		common.RenderError(w, r, renderRoleMutationError(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deleteRole handles deleting a role
-func (rs *Resource) deleteRole(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeleteRole(r.Context(), id); err != nil {
-		if common.IsConstraintViolation(err) {
-			common.RenderError(w, r, common.ErrorConflictMessage("Rolle kann nicht gelöscht werden: Rolle ist aktuell Konten zugewiesen"))
-			return
-		}
-		common.RenderError(w, r, renderRoleMutationError(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// renderRoleMutationError maps service-layer role errors to appropriate HTTP responses.
-func renderRoleMutationError(err error) render.Renderer {
-	var authErr *authService.AuthError
-	if errors.As(err, &authErr) {
-		if errors.Is(authErr.Err, authService.ErrSystemRoleImmutable) {
-			return common.ErrorForbidden(authErr.Err)
-		}
-		if errors.Is(authErr.Err, authService.ErrRoleNotFound) {
-			return common.ErrorNotFound(authErr.Err)
-		}
-	}
-	// FindByID failures (sql.ErrNoRows wrapped in DatabaseError) → 404
-	if errors.Is(err, sql.ErrNoRows) {
-		return common.ErrorNotFound(errors.New("role not found"))
-	}
-	return common.ErrorInternalServer(err)
-}
-
-// listRoles handles listing roles
-func (rs *Resource) listRoles(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if name := r.URL.Query().Get("name"); name != "" {
-		filters["name"] = name
-	}
-
-	roles, err := rs.AuthService.ListRoles(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*RoleResponse, 0, len(roles))
-	for _, role := range roles {
-		responses = append(responses, toRoleResponse(role))
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Roles retrieved successfully")
-}
-
-// getAccountRoles handles getting roles for an account
-func (rs *Resource) getAccountRoles(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	roles, err := rs.AuthService.GetAccountRoles(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*RoleResponse, 0, len(roles))
-	for _, role := range roles {
-		responses = append(responses, toRoleResponse(role))
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account roles retrieved successfully")
-}
-
-func (rs *Resource) getCaregiverCapability(w http.ResponseWriter, r *http.Request) {
-	if rs.CaregiverCapabilityService == nil {
-		common.RenderError(w, r, common.ErrorInternalServer(errors.New("caregiver capability service is not configured")))
-		return
-	}
-
-	accountID, ok := common.ParseInt64IDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	state, err := rs.CaregiverCapabilityService.GetCaregiverCapability(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, caregiverCapabilityErrorRenderer(err))
-		return
-	}
-
-	common.Respond(w, r, http.StatusOK, state, "Caregiver capability retrieved successfully")
-}
-
-func (rs *Resource) enableCaregiverCapability(w http.ResponseWriter, r *http.Request) {
-	if rs.CaregiverCapabilityService == nil {
-		common.RenderError(w, r, common.ErrorInternalServer(errors.New("caregiver capability service is not configured")))
-		return
-	}
-
-	accountID, ok := common.ParseInt64IDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	req := &EnableCaregiverCapabilityRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	state, err := rs.CaregiverCapabilityService.EnableCaregiverCapability(r.Context(), accountID, userModel.EnableCaregiverCapabilityInput{
-		FirstName: req.FirstName,
-		LastName:  req.LastName,
-		Position:  req.Position,
-	})
-	if err != nil {
-		common.RenderError(w, r, caregiverCapabilityErrorRenderer(err))
-		return
-	}
-
-	common.Respond(w, r, http.StatusOK, state, "Caregiver capability enabled successfully")
-}
-
-func (rs *Resource) disableCaregiverCapability(w http.ResponseWriter, r *http.Request) {
-	if rs.CaregiverCapabilityService == nil {
-		common.RenderError(w, r, common.ErrorInternalServer(errors.New("caregiver capability service is not configured")))
-		return
-	}
-
-	accountID, ok := common.ParseInt64IDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	state, err := rs.CaregiverCapabilityService.DisableCaregiverCapability(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, caregiverCapabilityErrorRenderer(err))
-		return
-	}
-
-	common.Respond(w, r, http.StatusOK, state, "Caregiver capability disabled successfully")
-}
-
-func caregiverCapabilityErrorRenderer(err error) render.Renderer {
-	var blockedErr *usersService.CaregiverCapabilityBlockedError
-	var accountTenantErr *usersService.AccountNotAssignedToTenantError
-	var usersErr *usersService.UsersError
-	var validationErr *usersService.ValidationError
-
-	switch {
-	case errors.As(err, &blockedErr):
-		return common.NewCaregiverCapabilityBlockedResponse(
-			http.StatusConflict,
-			blockedErr.Error(),
-			blockedErr.Reasons,
-		)
-	case errors.Is(err, authService.ErrAccountNotFound), errors.As(err, &accountTenantErr):
-		return common.ErrorNotFound(errors.New("account not found"))
-	case errors.As(err, &usersErr) && errors.As(err, &validationErr):
-		return common.ErrorInvalidRequest(validationErr)
-	case errors.As(err, &usersErr):
-		return common.ErrorInternalServer(usersErr.Err)
-	default:
-		return common.ErrorInternalServer(err)
-	}
-}
-
-// Permission Management Endpoints
-
-// createPermission handles creating a new permission
-func (rs *Resource) createPermission(w http.ResponseWriter, r *http.Request) {
-	req := &CreatePermissionRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	permission, err := rs.AuthService.CreatePermission(r.Context(), req.Name, req.Description, req.Resource, req.Action)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	resp := &PermissionResponse{
-		ID:          permission.ID,
-		Name:        permission.Name,
-		Description: permission.Description,
-		Resource:    permission.Resource,
-		Action:      permission.Action,
-		CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-	}
-
-	common.Respond(w, r, http.StatusCreated, resp, "Permission created successfully")
-}
-
-// getPermissionByID handles getting a permission by ID
-func (rs *Resource) getPermissionByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	permission, err := rs.AuthService.GetPermissionByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("permission not found")))
-		return
-	}
-
-	resp := &PermissionResponse{
-		ID:          permission.ID,
-		Name:        permission.Name,
-		Description: permission.Description,
-		Resource:    permission.Resource,
-		Action:      permission.Action,
-		CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-		UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Permission retrieved successfully")
-}
-
-// updatePermission handles updating a permission
-func (rs *Resource) updatePermission(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	req := &UpdatePermissionRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	permission, err := rs.AuthService.GetPermissionByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("permission not found")))
-		return
-	}
-
-	permission.Name = req.Name
-	permission.Description = req.Description
-	permission.Resource = req.Resource
-	permission.Action = req.Action
-
-	if err := rs.AuthService.UpdatePermission(r.Context(), permission); err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// deletePermission handles deleting a permission
-func (rs *Resource) deletePermission(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidPermissionID)
-	if !ok {
-		return
-	}
-
-	if err := rs.AuthService.DeletePermission(r.Context(), id); err != nil {
-		if common.IsConstraintViolation(err) {
-			common.RenderError(w, r, common.ErrorConflictMessage("Berechtigung kann nicht gelöscht werden: Berechtigung ist aktuell Rollen oder Konten zugewiesen"))
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listPermissions handles listing permissions
-func (rs *Resource) listPermissions(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if resource := r.URL.Query().Get("resource"); resource != "" {
-		filters["resource"] = resource
-	}
-
-	if action := r.URL.Query().Get("action"); action != "" {
-		filters["action"] = action
-	}
-
-	permissions, err := rs.AuthService.ListPermissions(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Permissions retrieved successfully")
-}
-
-// getAccountPermissions handles getting permissions for an account
-func (rs *Resource) getAccountPermissions(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetAccountPermissions(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account permissions retrieved successfully")
-}
-
-// getAccountDirectPermissions handles getting only direct permissions for an account (not role-based)
-func (rs *Resource) getAccountDirectPermissions(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetAccountDirectPermissions(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Account direct permissions retrieved successfully")
-}
-
-// getRolePermissions handles getting permissions for a role
-func (rs *Resource) getRolePermissions(w http.ResponseWriter, r *http.Request) {
-	roleID, ok := common.ParseIntIDWithError(w, r, "roleId", common.MsgInvalidRoleID)
-	if !ok {
-		return
-	}
-
-	permissions, err := rs.AuthService.GetRolePermissions(r.Context(), roleID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*PermissionResponse, 0, len(permissions))
-	for _, permission := range permissions {
-		resp := &PermissionResponse{
-			ID:          permission.ID,
-			Name:        permission.Name,
-			Description: permission.Description,
-			Resource:    permission.Resource,
-			Action:      permission.Action,
-			CreatedAt:   permission.CreatedAt.Format(time.RFC3339),
-			UpdatedAt:   permission.UpdatedAt.Format(time.RFC3339),
-		}
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Role permissions retrieved successfully")
-}
-
-// Account Management Extension Endpoints
-
-// updateAccount handles updating an account
-func (rs *Resource) updateAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	account, err := rs.AuthService.GetAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("account not found")))
-		return
-	}
-
-	account.Email = req.Email
-	if req.Username != "" {
-		username := req.Username
-		account.Username = &username
-	}
-
-	if err := rs.AuthService.UpdateAccount(r.Context(), account); err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listAccounts handles listing accounts
-func (rs *Resource) listAccounts(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if email := r.URL.Query().Get("email"); email != "" {
-		filters["email"] = email
-	}
-
-	if active := r.URL.Query().Get("active"); active != "" {
-		switch active {
-		case "true":
-			filters["active"] = true
-		case "false":
-			filters["active"] = false
-		}
-	}
-
-	accounts, err := rs.AuthService.ListAccounts(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*AccountResponse, 0, len(accounts))
-	for _, account := range accounts {
-		resp := &AccountResponse{
-			ID:     account.ID,
-			Email:  account.Email,
-			Active: account.Active,
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Accounts retrieved successfully")
-}
-
-// getAccountsByRole handles getting accounts by role
-func (rs *Resource) getAccountsByRole(w http.ResponseWriter, r *http.Request) {
-	roleName := chi.URLParam(r, "roleName")
-
-	accounts, err := rs.AuthService.GetAccountsByRole(r.Context(), roleName)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*AccountResponse, 0, len(accounts))
-	for _, account := range accounts {
-		resp := &AccountResponse{
-			ID:     account.ID,
-			Email:  account.Email,
-			Active: account.Active,
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Accounts retrieved successfully")
-}
-
-// Password Reset Endpoints
-
-// initiatePasswordReset handles initiating a password reset
-func (rs *Resource) initiatePasswordReset(w http.ResponseWriter, r *http.Request) {
-	req := &PasswordResetRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	// Always return success to avoid revealing whether email exists, but handle rate limiting
-	_, err := rs.AuthService.InitiatePasswordReset(r.Context(), req.Email)
-	if err != nil {
-		var rateErr *authService.RateLimitError
-		if errors.As(err, &rateErr) {
-			// Prefer Retry-After seconds, fallback to RFC1123 format
-			retryAfterSeconds := rateErr.RetryAfterSeconds(time.Now())
-			if retryAfterSeconds > 0 {
-				w.Header().Set("Retry-After", strconv.Itoa(retryAfterSeconds))
-			} else if !rateErr.RetryAt.IsZero() {
-				w.Header().Set("Retry-After", rateErr.RetryAt.UTC().Format(http.TimeFormat))
-			}
-
-			common.RenderError(w, r, common.ErrorTooManyRequests(authService.ErrRateLimitExceeded))
-			return
-		}
-
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	slog.Default().Info("Password reset initiated", slog.String("email", req.Email))
-
-	common.Respond(w, r, http.StatusOK, nil, "If the email exists, a password reset link has been sent")
-}
-
-// resetPassword handles resetting password with token
-func (rs *Resource) resetPassword(w http.ResponseWriter, r *http.Request) {
-	req := &PasswordResetConfirmRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	if err := rs.AuthService.ResetPassword(r.Context(), req.Token, req.NewPassword); err != nil {
-		slog.Default().WarnContext(r.Context(), "Password reset failed", slog.String("error", err.Error()))
-
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrInvalidToken),
-				errors.Is(authErr.Err, sql.ErrNoRows):
-				// Both cases indicate the token is invalid or not found
-				common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("invalid or expired reset token")))
-				return
-			case errors.Is(authErr.Err, authService.ErrPasswordTooWeak):
-				common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrPasswordTooWeak))
-				return
-			}
-		}
-
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	slog.Default().Info("Password reset completed successfully")
-
-	common.Respond(w, r, http.StatusOK, nil, "Password reset successfully")
-}
-
-// Token Management Endpoints
-
-// cleanupExpiredTokens handles cleanup of expired tokens
-func (rs *Resource) cleanupExpiredTokens(w http.ResponseWriter, r *http.Request) {
-	count, err := rs.AuthService.CleanupExpiredTokens(r.Context())
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	response := map[string]int{"cleaned_tokens": count}
-	common.Respond(w, r, http.StatusOK, response, "Expired tokens cleaned up successfully")
-}
-
-// getActiveTokens handles getting active tokens for an account
-func (rs *Resource) getActiveTokens(w http.ResponseWriter, r *http.Request) {
-	accountID, ok := common.ParseIntIDWithError(w, r, "accountId", common.MsgInvalidAccountID)
-	if !ok {
-		return
-	}
-
-	tokens, err := rs.AuthService.GetActiveTokens(r.Context(), accountID)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	type TokenResponse struct {
-		ID         int64  `json:"id"`
-		Token      string `json:"token"`
-		Expiry     string `json:"expiry"`
-		Mobile     bool   `json:"mobile"`
-		Identifier string `json:"identifier,omitempty"`
-		CreatedAt  string `json:"created_at"`
-	}
-
-	responses := make([]*TokenResponse, 0, len(tokens))
-	for _, token := range tokens {
-		resp := &TokenResponse{
-			ID:        token.ID,
-			Token:     token.Token,
-			Expiry:    token.Expiry.Format(time.RFC3339),
-			Mobile:    token.Mobile,
-			CreatedAt: token.CreatedAt.Format(time.RFC3339),
-		}
-
-		if token.Identifier != nil {
-			resp.Identifier = *token.Identifier
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Active tokens retrieved successfully")
-}
-
-// Parent Account Management Endpoints
-
-// createParentAccount handles creating a parent account
-func (rs *Resource) createParentAccount(w http.ResponseWriter, r *http.Request) {
-	req := &CreateParentAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	parentAccount, err := rs.AuthService.CreateParentAccount(r.Context(), req.Email, req.Username, req.Password)
-	if err != nil {
-		var authErr *authService.AuthError
-		if errors.As(err, &authErr) {
-			switch {
-			case errors.Is(authErr.Err, authService.ErrEmailAlreadyExists):
-				common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrEmailAlreadyExists))
-			case errors.Is(authErr.Err, authService.ErrUsernameAlreadyExists):
-				common.RenderError(w, r, common.ErrorInvalidRequest(authService.ErrUsernameAlreadyExists))
-			default:
-				common.RenderError(w, r, common.ErrorInternalServer(err))
-			}
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	resp := &ParentAccountResponse{
-		ID:        parentAccount.ID,
-		Email:     parentAccount.Email,
-		Active:    parentAccount.Active,
-		CreatedAt: parentAccount.CreatedAt.Format(time.RFC3339),
-		UpdatedAt: parentAccount.UpdatedAt.Format(time.RFC3339),
-	}
-
-	if parentAccount.Username != nil {
-		resp.Username = *parentAccount.Username
-	}
-
-	common.Respond(w, r, http.StatusCreated, resp, "Parent account created successfully")
-}
-
-// getParentAccountByID handles getting a parent account by ID
-func (rs *Resource) getParentAccountByID(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	parentAccount, err := rs.AuthService.GetParentAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("parent account not found")))
-		return
-	}
-
-	resp := &ParentAccountResponse{
-		ID:        parentAccount.ID,
-		Email:     parentAccount.Email,
-		Active:    parentAccount.Active,
-		CreatedAt: parentAccount.CreatedAt.Format(time.RFC3339),
-		UpdatedAt: parentAccount.UpdatedAt.Format(time.RFC3339),
-	}
-
-	if parentAccount.Username != nil {
-		resp.Username = *parentAccount.Username
-	}
-
-	common.Respond(w, r, http.StatusOK, resp, "Parent account retrieved successfully")
-}
-
-// updateParentAccount handles updating a parent account
-func (rs *Resource) updateParentAccount(w http.ResponseWriter, r *http.Request) {
-	id, ok := common.ParseIntIDWithError(w, r, "id", common.MsgInvalidParentAccountID)
-	if !ok {
-		return
-	}
-
-	req := &UpdateAccountRequest{}
-	if err := render.Bind(r, req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(err))
-		return
-	}
-
-	parentAccount, err := rs.AuthService.GetParentAccountByID(r.Context(), id)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("parent account not found")))
-		return
-	}
-
-	parentAccount.Email = req.Email
-	if req.Username != "" {
-		username := req.Username
-		parentAccount.Username = &username
-	}
-
-	if err := rs.AuthService.UpdateParentAccount(r.Context(), parentAccount); err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	common.RespondNoContent(w, r)
-}
-
-// listParentAccounts handles listing parent accounts
-func (rs *Resource) listParentAccounts(w http.ResponseWriter, r *http.Request) {
-	// Parse query parameters for filtering
-	filters := make(map[string]interface{})
-
-	if email := r.URL.Query().Get("email"); email != "" {
-		filters["email"] = email
-	}
-
-	if active := r.URL.Query().Get("active"); active != "" {
-		switch active {
-		case "true":
-			filters["active"] = true
-		case "false":
-			filters["active"] = false
-		}
-	}
-
-	parentAccounts, err := rs.AuthService.ListParentAccounts(r.Context(), filters)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServer(err))
-		return
-	}
-
-	responses := make([]*ParentAccountResponse, 0, len(parentAccounts))
-	for _, account := range parentAccounts {
-		resp := &ParentAccountResponse{
-			ID:        account.ID,
-			Email:     account.Email,
-			Active:    account.Active,
-			CreatedAt: account.CreatedAt.Format(time.RFC3339),
-			UpdatedAt: account.UpdatedAt.Format(time.RFC3339),
-		}
-
-		if account.Username != nil {
-			resp.Username = *account.Username
-		}
-
-		responses = append(responses, resp)
-	}
-
-	common.Respond(w, r, http.StatusOK, responses, "Parent accounts retrieved successfully")
-}
-
-// getClientIP extracts the real client IP address from the request
-func getClientIP(r *http.Request) string {
-	return clientip.GetClientIPString(r)
 }
